Modified TransactionService
Added new methods.

+++ Backend/server/src/Service/TransactionService.php
@@ -1,6 +1,8 @@
 <?php
 namespace App\Service;
 
+use App\Entity\User;
+use App\Entity\Product;
 use App\Entity\Transaction;
 use Doctrine\ORM\EntityManagerInterface;
 
@@ -19,10 +21,42 @@ class TransactionService
         $this->em->flush();
     }
 
-    public function deleteTransaction(Transaction $transaction)
+    public function addTransactionByDetails(User $user, Product $product, string $productName,
+                                            int $price, bool $isCompleted, int $quantity)
     {
-        $this->em->remove($transaction);
-        $this->em->flush();
+        $transaction = new Transaction();
+        $transaction->setUsers($user)
+            ->setProduct($product)
+            ->setProductName($productName)
+            ->setPrice($price)
+            ->setIsCompleted($isCompleted)
+            ->setQuantity($quantity);
+        $this->addTransaction($transaction);
+    }
+
+    public function deleteTransaction(Transaction $transaction): bool
+    {
+        if ($this->checkIfTransactionExistsById($transaction->getId()))
+        {
+            $this->em->remove($transaction);
+            $this->em->flush();
+            return true;
+        }
+        return false;
+    }
+
+    public function deleteTransactionById(int $id): bool
+    {
+        $user = $this->getTransactionById($id);
+        if ($user == null) return false;
+        $this->deleteTransaction($user);
+        return true;
+    }
+
+    public function checkIfTransactionExistsById(int $id): bool
+    {
+        $transaction = $this->getTransactionById($id);
+        return $transaction == null ? false : true;
     }
 
     public function getTransactionById(int $id): ?Transaction
@@ -30,12 +64,14 @@ class TransactionService
         return $this->em->getRepository(Transaction::class)->find($id);
     }
 
-    public function getTransactions(int $limit)
+    public function getTransactions(int $limit, bool $isCompleted=null, int $userId=null): array
     {
-        return $this->em->getRepository(Transaction::class)->findBy(null, null, $limit);
+        $criteria = array();
+        if ($userId != null && $isCompleted != null) $criteria = array('users' => $userId, 'is_completed' => $isCompleted);
+        return $this->em->getRepository(Transaction::class)->findBy($criteria, null, $limit);
     }
 
-    public function getAllTransactions()
+    public function getAllTransactions(): array
     {
         return $this->em->getRepository(Transaction::class)->findAll();
     }
